Organize step one, two and three
