+++ src/app/user/Stake/Steps/StepThree.tsx
@@ -1,25 +1,80 @@
-import { StakeStatus } from '@/app/user/Stake/StakeStatus'
-import { StepProps } from '@/app/user/Stake/types'
 import { useStakingContext } from '@/app/user/Stake/StakingContext'
-import { useRouter } from 'next/navigation'
+import { StepProps } from '@/app/user/Stake/types'
+import { isUserRejectedTxError } from '@/components/ErrorPage/commonErrors'
+import { config } from '@/config'
+import { showToast } from '@/shared/lib/toastUtils'
+import { TX_MESSAGES } from '@/shared/txMessages'
+import { waitForTransactionReceipt } from 'wagmi/actions'
+import { Divider } from '../components/Divider'
+import { StepActionButtons } from '../components/StepActionButtons'
+import { StepWrapper } from '../components/StepWrapper'
+import { TokenAmountDisplay } from '../components/TokenAmountDisplay'
+import { TransactionStatus } from '../components/TransactionStatus'
+import { useStakeRIF } from '../hooks/useStakeRIF'
 
 export const StepThree = ({ onCloseModal = () => {} }: StepProps) => {
-  const { stakeTxHash, amountDataToReceive, tokenToReceive, actionName } = useStakingContext()
-  const router = useRouter()
-  if (!stakeTxHash) return null
+  const { amount, tokenToReceive, stakePreviewFrom: from, stakePreviewTo: to } = useStakingContext()
 
-  const handleCloseModal = () => {
-    router.push('/user')
-    onCloseModal()
+  const { onRequestStake, isRequesting, isTxPending, isTxFailed, stakeHash } = useStakeRIF(
+    amount,
+    tokenToReceive.contract,
+  )
+
+  const handleConfirmStake = async () => {
+    try {
+      const txHash = await onRequestStake()
+      await waitForTransactionReceipt(config, {
+        hash: txHash,
+      })
+      onCloseModal()
+    } catch (err: any) {
+      if (!isUserRejectedTxError(err)) {
+        showToast(TX_MESSAGES.staking.error)
+      }
+    }
   }
+
   return (
-    <StakeStatus
-      onReturnToBalances={handleCloseModal}
-      hash={stakeTxHash}
-      symbol={tokenToReceive.symbol}
-      amountReceived={amountDataToReceive.amountToReceive}
-      amountReceivedCurrency={amountDataToReceive.amountToReceiveConvertedToCurrency}
-      actionName={actionName}
-    />
+    <StepWrapper
+      currentStep={3}
+      progress={100}
+      description="Make sure that everything is correct before continuing:"
+    >
+      <div className="flex flex-col md:flex-row md:items-start md:justify-between mb-8">
+        <TokenAmountDisplay
+          label="From"
+          amount={amount}
+          tokenSymbol={from.tokenSymbol}
+          amountInCurrency={from.amountConvertedToCurrency}
+          balance={from.balance}
+        />
+        <TokenAmountDisplay
+          label="To"
+          amount={amount}
+          tokenSymbol={to.tokenSymbol}
+          amountInCurrency={to.amountConvertedToCurrency}
+          balance={to.balance}
+          isFlexEnd
+        />
+      </div>
+
+      <TransactionStatus txHash={stakeHash} isTxFailed={isTxFailed} failureMessage="Stake TX failed." />
+
+      <Divider />
+
+      <StepActionButtons
+        primaryButton={{
+          label: isRequesting ? 'Requesting...' : 'Confirm stake',
+          onClick: handleConfirmStake,
+          disabled: !amount || Number(amount) <= 0,
+        }}
+        secondaryButton={{
+          label: 'Cancel',
+          onClick: onCloseModal,
+        }}
+        isTxPending={isTxPending}
+        isRequesting={isRequesting}
+      />
+    </StepWrapper>
   )
 }
